+++ examples/yesod-with-metrics/app/Main.hs
@@ -16,6 +16,7 @@ import qualified Data.Text as T
 import           Foundation
 import           Network.Wai.Handler.Warp (run)
 import           System.Metrics           (newStore, registerGcMetrics, sampleAll)
+import           System.Random            (randomRIO)
 import           Yesod
 import qualified Yesod.Routes.Metrics as Yesod
 import           System.Remote.Monitoring (forkServerWith)
@@ -27,6 +28,12 @@ routesFileContents = $(embedFile "config/routes")
 
 getHomeR :: Handler Text
 getHomeR = return "Hello World!"
+
+getDelayR :: Int -> Int -> Handler Text
+getDelayR start stop = do
+  waitTime <- liftIO $ randomRIO (start,stop)
+  liftIO $ threadDelay waitTime -- uses microseconds
+  return $ T.pack . show $ waitTime
 
 postNewUserR :: Handler Text
 postNewUserR = return $ "Making a new user"